[3769] Added env var,KEA_PIDFILE_DIR; D2 now uses a PIDFile
src/lib/dhcpsrv/daemon.c/h
    Daemon::Daemon() - Constructor will now override the default PID
    directory with the value of env variable KEA_PIDFILE_DIR.  This
    provides a simple means to alter the value for tests.

    Added am_file_author_ flag so Daemon instances will only delete
    a file they have written.

src/lib/testutils/dhcp_test_lib.sh.in
    - verify_server_pid() - new function which verifies that a server has a PID file
    and that it contains the server's PID, and that the process is alive.

src/bin/keactrl/tests/Makefile.am
    - added export of KEA_PIDFILE_DIR to override default PID directory during tests

Added PID file creation to D2
    src/bin/d2/d_controller.cc
    - DControllerBase::launch() - Added block to createPIDFile()

    -DControllerBase::parseArgs() Replaced call to Daemon::init()
    with call to Daemon::setConfigFile()

    src/bin/d2/tests/Makefile.am
    - added export of KEA_PIDFILE_DIR to override default PID directory during tests

    src/bin/d2/tests/d2_process_tests.sh.in
    - dupcliate_server_start_test() - new test which verifies that D2 cannot be
    started twice (with the same configuration file)

    src/bin/d2/tests/d2_unittests.cc
    - main(int argc, char* argv[]) sets environment variable KEA_PIDFILE_DIR
     to override default PID diretory during tests

src/lib/util/pid_file.cc/h
src/lib/util/tests/pid_file_unittest.cc
    Changed PIDFile::check() to return either the PID contained in the PID file
    if the process is alive, or 0, rather than bool.  This permits callers
    to see/log the PID.

+++ src/lib/dhcpsrv/daemon.cc
@@ -39,11 +39,18 @@ std::string Daemon::config_file_ = "";
 
 Daemon::Daemon()
     : signal_set_(), signal_handler_(), proc_name_(""),
-    pid_file_dir_(DHCP_DATA_DIR), pid_file_() {
+    pid_file_dir_(DHCP_DATA_DIR), pid_file_(), am_file_author_(false) {
+
+    // The pid_file_dir can be overridden via environment variable
+    // This is primarily intended to simplify testing
+    const char* const env = getenv("KEA_PIDFILE_DIR");
+    if (env) {
+        pid_file_dir_ = env;
+    }
 }
 
 Daemon::~Daemon() {
-    if (pid_file_) {
+    if (pid_file_ && am_file_author_) {
         pid_file_->deleteFile();
     }
 }
@@ -191,9 +198,10 @@ Daemon::createPIDFile(int pid) {
     }
 
     // If we find a pre-existing file containing a live PID we bail.
-    if (pid_file_->check()) {
-        isc_throw(DaemonPIDExists, "Daemon::createPIDFile " << proc_name_
-                  << " already running?, PID file: " << getPIDFileName());
+    int chk_pid = pid_file_->check();
+    if (chk_pid > 0) {
+        isc_throw(DaemonPIDExists, "Daemon::createPIDFile: PID: " << chk_pid
+                 << " exists, PID file: " << getPIDFileName());
     }
 
     if (pid == 0) {
@@ -203,6 +211,8 @@ Daemon::createPIDFile(int pid) {
         // Write the PID we were given
         pid_file_->write(pid);
     }
+
+    am_file_author_ = true;
 }
 
 };
